+++ src/main/kotlin/net/pwall/text/URIStringMapper.kt
@@ -25,21 +25,28 @@
 
 package net.pwall.text
 
+import net.pwall.text.StringMapper.buildResult
+import net.pwall.text.StringMapper.fromHexDigit
 import net.pwall.text.StringMapper.mapCharacters
-import net.pwall.text.StringMapper.mapSubstring
+import net.pwall.text.StringMapper.mapSubstrings
 import net.pwall.util.IntOutput.append2Hex
 
 /**
- * String mapping functions to perform URI percent-encoding and decoding.
+ * String mapping functions to perform URI percent-encoding and -decoding.
  *
  * @author  Peter Wall
  */
 object URIStringMapper {
 
+    private val spaceResult = CharMapResult(1, ' ')
+
+    private const val illegalMessage = "Illegal URI escape sequence"
+    private const val incompleteMessage = "Incomplete URI escape sequence"
+
     /**
      * Encode string using URI percent-encoding.
      */
-    fun String.encodeURI() = mapCharacters {
+    fun String.encodeURI(): String = mapCharacters {
         if (it.isUnreservedForURI() || it == '$') null else StringBuilder(3).apply {
             append('%')
             append2Hex(this, it.code)
@@ -47,22 +54,27 @@ object URIStringMapper {
     }
 
     /**
-     * Decode string from URI percent-encoding.
+     * Decode string from URI percent-encoding (also decodes "+" as space, despite the fact that the above function does
+     * not encode space this way).
      */
-    fun String.decodeURI() = mapSubstring {
-        if (this[it] != '%') null else StringMapper.buildResult(this, it, 3) {
-            (this[it + 1].fromHexDigit() shl 4) or this[it + 2].fromHexDigit()
+    fun String.decodeURI(): String = mapSubstrings {
+        when (this[it]) {
+            '%' -> buildResult(this, it, 3, incompleteMessage) {
+                try {
+                    (this[it + 1].fromHexDigit() shl 4) or this[it + 2].fromHexDigit()
+                } catch (_: NumberFormatException) {
+                    throw IllegalArgumentException(illegalMessage)
+                }
+            }
+            '+' -> spaceResult
+            else -> null
         }
     }
 
+    /**
+     * Test whether a character is in the "unreserved for URI" set.
+     */
     fun Char.isUnreservedForURI(): Boolean = this in 'A'..'Z' || this in 'a'..'z' || this in '0'..'9' ||
             this == '-' || this == '.' || this == '_' || this == '~'
-
-    fun Char.fromHexDigit(): Int = when (this) {
-        in '0'..'9' -> this.code - '0'.code
-        in 'A'..'F' -> this.code - 'A'.code + 10
-        in 'a'..'f' -> this.code - 'a'.code + 10
-        else -> throw IllegalArgumentException("Illegal URI escape sequence")
-    }
 
 }