+++ src/test/kotlin/net/pwall/text/UTF8StringMapperTest.kt
@@ -29,29 +29,40 @@ import kotlin.test.Test
 import kotlin.test.assertFailsWith
 import kotlin.test.expect
 
-import net.pwall.text.UTF8StringMapper.fromUTF8
-import net.pwall.text.UTF8StringMapper.toUTF8
+import net.pwall.text.UTF8StringMapper.decodeUTF8
+import net.pwall.text.UTF8StringMapper.encodeUTF8
 
 class UTF8StringMapperTest {
 
     @Test fun `should map to UTF-8`() {
-        expect("") { "".toUTF8() }
-        expect("[ \u00C2\u00A9 \u00C3\u00B7 ]") { "[ \u00A9 \u00F7 ]".toUTF8() }
-        expect("[ \u00E2\u0080\u0094 ]") { "[ \u2014 ]".toUTF8() }
+        expect("") { "".encodeUTF8() }
+        expect("[ \u00C2\u00A9 \u00C3\u00B7 ]") { "[ \u00A9 \u00F7 ]".encodeUTF8() }
+        expect("[ \u00E2\u0080\u0094 ]") { "[ \u2014 ]".encodeUTF8() }
+        expect("[ \u00EF\u00BB\u00BF ]") { "[ \uFEFF ]".encodeUTF8() }
     }
 
     @Test fun `should map from UTF-8`() {
-        expect("") { "".fromUTF8() }
-        expect("[ \u00A9 \u00F7 ]") { "[ \u00C2\u00A9 \u00C3\u00B7 ]".fromUTF8() }
-        expect("[ \u2014 ]") { "[ \u00E2\u0080\u0094 ]".fromUTF8() }
+        expect("") { "".decodeUTF8() }
+        expect("[ \u00A9 \u00F7 ]") { "[ \u00C2\u00A9 \u00C3\u00B7 ]".decodeUTF8() }
+        expect("[ \u2014 ]") { "[ \u00E2\u0080\u0094 ]".decodeUTF8() }
+        expect("[ \uFEFF ]") { "[ \u00EF\u00BB\u00BF ]".decodeUTF8() }
     }
 
     @Test fun `should fail on incomplete UTF-8 sequence`() {
-        assertFailsWith<IllegalArgumentException> { " \u00C2".fromUTF8() }.let {
+        assertFailsWith<IllegalArgumentException> { " \u00C2".decodeUTF8() }.let {
             expect("Incomplete UTF-8 sequence") { it.message }
         }
-        assertFailsWith<IllegalArgumentException> { " \u00E2\u0080".fromUTF8() }.let {
+        assertFailsWith<IllegalArgumentException> { " \u00E2\u0080".decodeUTF8() }.let {
             expect("Incomplete UTF-8 sequence") { it.message }
+        }
+    }
+
+    @Test fun `should fail on illegal UTF-8 sequence`() {
+        assertFailsWith<IllegalArgumentException> { " \u00C2\u0000".decodeUTF8() }.let {
+            expect("Illegal UTF-8 sequence") { it.message }
+        }
+        assertFailsWith<IllegalArgumentException> { " \u00E2\u0080\u0030".decodeUTF8() }.let {
+            expect("Illegal UTF-8 sequence") { it.message }
         }
     }