+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [2.0] - 2022-10-23
+### Added
+- `JSONStringMapper`: JSON string encoding and decoding
+### Changed
+- `StringMapper`, `UTF8StringMapper`: renamed functions (for consistency)
+
 ## [1.1] - 2022-10-16
 ### Changed
 - `StringMapper`, `URIStringMapper`, `UTF8StringMapper`: simplified API for `mapSubstring`